# [Vulkan] Add the 2D case to Layernorm operator